Allow to customize the fallback route when no RelayState is provided
If the SAMLResponse doesn't contain a RelayState parameter (a previous
page that we want to redirect the user to) right now the plugin
redirects to the user dashboard (the user.me route).

This change keeps this default behaviour but allows to customize the
route in case fallback redirection to another page is required.